Homework, Java-2: lesson 2, exceptions Доработка основных методов.
