+++ src/Java_2/j2_lesson2/MainClass.java
@@ -3,7 +3,7 @@ package Java_2.j2_lesson2;
 public class MainClass {
 
     public static void main(String[] args) {
-        ProcessMatrixCreation.createStrMatrix();
+        ProcessMatrix.processMatrix();
     }
 
 }
